Large Commit (chams, doubletap, etc) Read Body
Added tickbase exploit with indicator
Added chams (using DrawModelExecute) (flat, texture, fresnel. With overlay chams)
Added debug flag for disabling friend related checks (temp fix for the Steam ID bug)
Thirdperson now collides with map geometry
Pickup respawn timers now center to the pickup and not the pickuper
Added injection script (debug.sh) that attaches GDB and makes it interactable
Set build mode to force C++23 standards
Added libCURL as a dependency (TODO: add to package scripts)
Added support for the PiShock API (debug)
Added Watermark toggle
Added custom Watermark text
More undocumented changes regarding hooked methods, dependencies, README changes, Makefile changes, build.sh changes, and general code changes

+++ classes/team_objective_resource.hpp
@@ -25,7 +25,7 @@ public:
   }
 
   bool can_team_capture(int index, enum tf_team team) {
-    int array_index = index + (team * MAX_CONTROL_POINTS);
+    int array_index = index + ((int)(team) * MAX_CONTROL_POINTS);
     return ((bool*)(this + 0xF74))[array_index];
   }
